requirements/production.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements/production.in
#
asgiref==3.6.0
    # via django
dj-rest-auth==3.0.0
    # via -r requirements/production.in
django==4.2
    # via
    #   -r requirements/production.in
    #   dj-rest-auth
    #   django-extensions
    #   django-filter
    #   djangorestframework
    #   drf-nested-routers
django-environ==0.10.0
    # via -r requirements/production.in
django-extensions==3.2.1
    # via -r requirements/production.in
django-filter==23.1
    # via -r requirements/production.in
djangorestframework==3.14.0
    # via
    #   -r requirements/production.in
    #   dj-rest-auth
    #   drf-nested-routers
drf-nested-routers==0.93.4
    # via -r requirements/production.in
markupsafe==2.1.2
    # via werkzeug
pytz==2023.3
    # via djangorestframework
sqlparse==0.4.4
    # via django
werkzeug==2.2.3
    # via -r requirements/production.in
